Add owner vehicle APIs: create, update and fetch vehicles using EF Core

+++ backend/Dot Net/RentIt_owner_services/Models/User.cs
@@ -35,7 +35,7 @@ public partial class User
 
     public string? Answer { get; set; }
 
-    public sbyte? IsActive { get; set; }
+    public ulong? IsActive { get; set; }
 
     public string? ApprovalStatus { get; set; }
